+++ src/index.js
@@ -2,7 +2,6 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import App from "./App";
-import reportWebVitals from "./reportWebVitals";
 import store from "./redux/config";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
@@ -11,8 +10,3 @@ root.render(
     <App />
   </Provider>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
