+++ src/shared/Router.js
@@ -2,12 +2,11 @@ import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import MainPage from "../pages/MainPage";
 import DetailPage from "../pages/DetailPage";
-import UploadPostPage from "../pages/UploadPostPage";
 import SignUpPage from "../pages/SignUpPage";
 import KakaoCallback from "../components/login/KakaoCallback";
 import SignInPage from "../pages/SignInPage";
 import EditPage from '../pages/EditPage'
-import UploadPageRefactored from '../pages/UploadPageRefactored'
+import UploadPostPage from '../pages/UploadPostPage'
 
 const Router = () => {
   return (
@@ -17,9 +16,8 @@ const Router = () => {
           <Route path="/" element={<MainPage />} />
           <Route path="/post/:postId" element={<DetailPage />} />
           <Route path="/sign-in" element={<SignInPage />} />
-          <Route path="/upload" element={<UploadPageRefactored />} />
+          <Route path="/upload" element={<UploadPostPage />} />
           <Route path="/edit/:postId" element={<EditPage />} />
-          <Route path="/update/:postId" element={<UploadPostPage />} />
           <Route path="/sign-up" element={<SignUpPage />} />
           <Route path="/kakaoCallback" element={<KakaoCallback />} />
         </Routes>
